Module Make.User

The User module provides access to User API.

val by_id : id:string -> unit -> Gitlab_t.user Response.t Monad.t

by_id ~id () is the user information for user id.

See Single User.

val by_name : name:string -> unit -> Gitlab_t.users Response.t Monad.t

by_name ~name () search for user by name.

See List Users.

val current_user : token:Token.t -> unit -> Gitlab_t.current_user Response.t Monad.t

current_user ~token () is the current user for token. See Current Authenticated User.

val projects : id:string -> unit -> Gitlab_t.projects_short Response.t Monad.t

projects ~id () list user projects for user id.

See List User Projects.

val merge_requests : token:Token.t -> ?state:Gitlab_t.state -> ?milestone:string -> ?labels:string list -> ?author:string -> ?author_username:string -> ?my_reaction:string -> ?scope:Gitlab_t.merge_request_scope -> unit -> Gitlab_t.merge_request Stream.t

merge_requests () list all merge requests the authenticated user has access to.

See List merge requests.

val issues : token:Token.t -> ?state:Gitlab_t.state -> ?assignee:string -> ?assignee_username:string -> ?milestone:string -> ?labels:string list -> ?author:string -> ?author_username:string -> ?scope:Gitlab_t.merge_request_scope -> unit -> Gitlab_t.issue Stream.t

issues ~token () returns a list of project issues.

See List issues.

val events : token:Token.t -> id:string -> ?action:Gitlab_t.event_action_type -> ?target_type:Gitlab_t.event_target_type -> ?per_page:int -> ?after:string -> ?sort:Gitlab_t.sort -> unit -> Gitlab_t.event Stream.t

events ~token ~id get the contribution events for the specified user.

See Get user contribution events.

module PersonalAccessToken : sig ... end

Personal access tokens for User authentication.