Module Group.Project

val by_name : ?token:Token.t -> owner:string -> name:string -> unit -> Gitlab_t.projects_short Response.t Monad.t

by_name ~group ~name () retrieves projects owned by group with a name like name. Depending on the name used this will return 1 or more matches. Supply a token to access private projects.

There is no direct fetch by name API in GitLab.