Module Project.Commit

Commit operates on a repository's commits.

val commits : ?token:Token.t -> project_id:int -> ?ref_name:string -> ?since:string -> ?until:string -> ?path:string -> ?all:bool -> unit -> Gitlab_t.commit Stream.t

commits ?token ~project_id () list all commits for a project.

See List repository commits.

val commit : ?token:Token.t -> project_id:int -> sha:string -> ?stats:bool -> unit -> Gitlab_t.commit Response.t Monad.t

commit ?token ~project_id ~sha get a specified commit by the commit hash or name of a branch or tag.

See Get a single commit.

val comments : ?token:Token.t -> project_id:int -> sha:string -> unit -> Gitlab_t.commit_comment Stream.t

comments ?token ~project_id ~sha get comments of a commit.

See Get the comments of a commit.

val comment : ?token:Token.t -> project_id:int -> sha:string -> Gitlab_t.new_comment -> unit -> Gitlab_t.commit_commented Response.t Monad.t

comment ~token ~project_id ~sha ~new_comment adds a comment to a commit.

See Post comment to a commit.

val statuses : ?token:Token.t -> project_id:int -> sha:string -> ?ref_name:string -> ?stage:string -> ?name:string -> ?all:bool -> unit -> Gitlab_t.commit_status Stream.t

statuses ~token ~project_id ~sha lists the statuses of a commit.

See List the statuses of a commit in a project.

val status : token:Token.t -> project_id:int -> sha:string -> Gitlab_t.new_status -> unit -> Gitlab_t.commit_status Response.t Monad.t

status ~token ~project_id ~sha ~state adds or updates the build status of a commit.

See Post the build status to a commit.