Module Notes.Merge_request

The Merge_request module provides access to Merge requests notes API.

val list : ?token:Token.t -> project_id:int -> merge_request_iid:string -> ?sort:Gitlab_t.sort -> unit -> Gitlab_t.note Stream.t

list ?token ~project_id ~merge_request_iid Request a list of a merge request notes. See List all merge request notes.

val by_id : ?token:Token.t -> project_id:int -> merge_request_iid:string -> note_id:int -> unit -> Gitlab_t.note Response.t Monad.t

by_id ?token ~project_id ~merge_request_iid ~note_id Get a single note for a given merge request. See Get single merge request note.

val create : token:Token.t -> project_id:int -> merge_request_iid:string -> create_note:Gitlab_t.create_note -> unit -> Gitlab_t.note Response.t Monad.t

create ?token ~project_id ~merge_request_iid ~create_note Creates a new note. See Create new merge request note.

val update : token:Token.t -> project_id:int -> merge_request_iid:string -> note_id:int -> body:string -> unit -> Gitlab_t.note Response.t Monad.t

update ?token ~project_id ~merge_request_iid ~note_id ~body Updates a note. See Modify existing note of a merge request.

val delete : token:Token.t -> project_id:int -> merge_request_iid:string -> note_id:int -> unit -> unit Response.t Monad.t

delete ?token ~project_id ~merge_request_iid ~note_id Deletes a note. See Delete a merge request note.