Module Gitlab_s

module type Gitlab = sig ... end

module type Env = sig ... end

A module of this type is required in order to construct a Gitlab module using Gitlab_core.Make.

module type Time = sig ... end

A module of this type is required in order to construct a Gitlab module using Gitlab_core.Make.